// src/pages/package-B/preBuy-list/types.ts
import Actions from './actions';
import {FollowListResponse} from 'api/GoodsFollowBaseController';

export interface IMainReducer {
  isReady: boolean;
  isLoading?: boolean;

  edit: IMainEdit;

  invalidFlag: IMainInvalidFlag;

  totalCount: IMainTotalCount;

  ifEvalShow: IMainIsShow;

  request: IMainRequest;

  total: IMainTotal;

  list: FollowListResponse;

  ifModify: false;

  selectedList: any;

  ifSelectAll: false;

  reload: boolean;

  serverTime: any;
}

export type ActionType = ReturnType<typeof Actions>;
export type IAllReducerProps = {
  main: IMainReducer;

  [name: string]: any;
};

//默认是全部的属性,可以自定义
export type IProps = IAllReducerProps & ActionType;

export type ICollectionBottomProps = {};
export type ICollectionBottomState = {};

export type ICollectionListItemProps = {};
export type ICollectionListItemState = {};

export type ICollectionListProps = {};
export type ICollectionListState = {};

export type ICollectionTopProps = {};
export type ICollectionTopState = {};

export type IMainEdit = boolean;
export type IMainInvalidFlag = boolean;
export type IMainTotalCount = number;
export type IMainIsShow = boolean;
export interface IMainRequest {
  q?: string;
  pageNum?: number;
  pageSize?: number;
  [k: string]: any;
}
export type IMainTotal = number;

//create by moon https://github.com/creasy2010/moon